requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
pymysql>=1.0.2
openai==1.72.0
python-dotenv>=1.0
google-generativeai==0.8.5
gunicorn==21.2.0
datetime